+++ examples/table_movie.py
@@ -6,7 +6,6 @@ from rich import box
 from rich.align import Align
 from rich.console import Console
 from rich.live import Live
-from rich.measure import Measurement
 from rich.table import Table
 from rich.text import Text